+++ src/components/Dropdown/SortDropdown/SortDropdown.tsx
@@ -2,9 +2,9 @@ import React, { ComponentProps, ComponentPropsWithRef } from 'react'
 import { type FC, type MouseEventHandler } from 'react'
 
 import { Button } from '../../Button'
+import { Fieldset } from '../../Fieldset'
 import { FormControl } from '../../FormControl'
 import { Cluster, Stack } from '../../Layout'
-import { Fieldset } from '../../NewFieldset'
 import { RadioButton } from '../../RadioButton'
 import { Select } from '../../Select'
 import { Dropdown } from '../Dropdown'
