+++ src/components/FormControl/index.ts
@@ -1,1 +1,1 @@
-export { FormControl } from './FormControl'
+export { ActualFormControl, FormControl } from './FormControl'
