--- src/components/FormGroup/index.ts
@@ -1,1 +1,0 @@
-export { FormGroup } from './FormGroup'
