+++ src/components/Layout/Cluster/Cluster.stories.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import styled, { css } from 'styled-components'
 
 import { Base } from '../../Base'
+import { Fieldset } from '../../Fieldset'
 import { Heading } from '../../Heading'
-import { Fieldset } from '../../NewFieldset'
 import { RadioButton } from '../../RadioButton'
 import { StatusLabel } from '../../StatusLabel'
 import { Stack } from '../Stack'
