--- src/components/NewFieldset/index.ts
@@ -1,1 +1,0 @@
-export { Fieldset } from './Fieldset'
