+++ src/components/NotificationBar/NotificationBar.stories.tsx
@@ -4,8 +4,8 @@ import styled, { css } from 'styled-components'
 
 import { Button } from '../Button'
 import { CheckBox } from '../CheckBox'
+import { Fieldset } from '../Fieldset'
 import { Cluster, Stack } from '../Layout'
-import { Fieldset } from '../NewFieldset'
 import { RadioButton } from '../RadioButton'
 import { Text } from '../Text'
 import { TextLink as shrTextLink } from '../TextLink'
