+++ src/components/RadioButton/VRTRadioButton.stories.tsx
@@ -2,8 +2,8 @@ import { StoryFn } from '@storybook/react'
 import React, { ChangeEvent, useState } from 'react'
 import styled from 'styled-components'
 
+import { Fieldset } from '../Fieldset'
 import { InformationPanel } from '../InformationPanel'
-import { Fieldset } from '../NewFieldset'
 
 import { All } from './RadioButton.stories'
